+++ handler/spa.go
@@ -4,10 +4,11 @@ import (
 	"bytes"
 	"encoding/json"
 	"io"
+	"mime"
 	"net/http"
 	"os"
 	"path/filepath"
-	"strings"
+	"strconv"
 	"sync"
 	"time"
 
@@ -30,6 +31,7 @@ type Spa struct {
 	bootstrapContent []byte
 	bootstrapModTime time.Time
 	bootstrapOnce    sync.Once
+	bootstrapCType   string
 	config           *config.Spa
 	modifier         []hcl.Body
 	srvOptions       *server.Options
@@ -77,38 +79,40 @@ func (s *Spa) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 	var content io.ReadSeeker
 	var modTime time.Time
 
-	if len(s.bootstrapContent) > 0 {
-		content = strings.NewReader(string(s.bootstrapContent))
-		modTime = s.bootstrapModTime
-	} else {
-		file, err := os.Open(s.config.BootstrapFile)
-		if err != nil {
-			if _, ok := err.(*os.PathError); ok {
-				s.srvOptions.ServerErrTpl.WithError(errors.RouteNotFound).ServeHTTP(rw, req)
-				return
-			}
-
-			s.srvOptions.ServerErrTpl.WithError(errors.Configuration).ServeHTTP(rw, req)
-			return
-		}
-		content = file
-		defer file.Close()
-
-		fileInfo, err := file.Stat()
-		if err != nil || fileInfo.IsDir() {
-			s.srvOptions.ServerErrTpl.WithError(errors.Configuration).ServeHTTP(rw, req)
-			return
-		}
-
-		modTime = fileInfo.ModTime()
-	}
-
 	evalContext := eval.ContextFromRequest(req)
-
 	if r, ok := rw.(*writer.Response); ok {
 		r.AddModifier(evalContext.HCLContext(), s.modifier...)
 	}
 
+	if l := len(s.bootstrapContent); l > 0 {
+		setLastModified(rw, s.bootstrapModTime)
+		rw.Header().Set("Content-Type", s.bootstrapCType)
+		rw.Header().Set("Content-Length", strconv.Itoa(l))
+		rw.WriteHeader(http.StatusOK) // required for ResponseWriter
+		_, _ = rw.Write(s.bootstrapContent)
+		return
+	}
+
+	file, err := os.Open(s.config.BootstrapFile)
+	if err != nil {
+		if _, ok := err.(*os.PathError); ok {
+			s.srvOptions.ServerErrTpl.WithError(errors.RouteNotFound).ServeHTTP(rw, req)
+			return
+		}
+
+		s.srvOptions.ServerErrTpl.WithError(errors.Configuration).ServeHTTP(rw, req)
+		return
+	}
+	content = file
+	defer file.Close()
+
+	fileInfo, err := file.Stat()
+	if err != nil || fileInfo.IsDir() {
+		s.srvOptions.ServerErrTpl.WithError(errors.Configuration).ServeHTTP(rw, req)
+		return
+	}
+
+	modTime = fileInfo.ModTime()
 	http.ServeContent(rw, req, s.config.BootstrapFile, modTime, content)
 }
 
@@ -152,6 +156,14 @@ func (s *Spa) replaceBootstrapData(ctx *hcl.EvalContext, reader io.ReadCloser) e
 	}
 	s.bootstrapContent = bytes.Replace(b, []byte(bootstrapName), escapedData.Bytes(), 1)
 
+	s.bootstrapCType = mime.TypeByExtension(filepath.Ext(s.config.BootstrapFile))
+	if s.bootstrapCType == "" {
+		// read a chunk to decide between utf-8 text and binary
+		var buf [512]byte
+		n, _ := io.ReadFull(bytes.NewBuffer(b), buf[:])
+		s.bootstrapCType = http.DetectContentType(buf[:n])
+	}
+
 	return nil
 }
 
@@ -161,4 +173,10 @@ func (s *Spa) Options() *server.Options {
 
 func (s *Spa) String() string {
 	return "spa"
+}
+
+func setLastModified(w http.ResponseWriter, modtime time.Time) {
+	if !modtime.IsZero() {
+		w.Header().Set("Last-Modified", modtime.UTC().Format(http.TimeFormat))
+	}
 }